# web/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jeevanti Hospitals Management System API",
    "description": "Comprehensive API documentation for the Jeevanti Hospitals Management System (HMS) built with Next.js, Prisma, MongoDB, and NextAuth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/patients": {
      "get": {
        "summary": "List patients",
        "description": "Retrieve a list of patients with search filtering on UHID, mobile, or name.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchType",
            "in": "query",
            "description": "Field to search by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "uhid",
                "mobile"
              ],
              "default": "name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patients retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create a new patient",
        "description": "Registers a new patient and generates a unique UHID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "409": {
            "description": "Conflict - Patient already exists with this mobile number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "summary": "Get patient details",
        "description": "Retrieve details of a specific patient including appointments, consultations, admissions, lab orders, and invoices.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Patient ID (MongoDB ObjectId)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDetails"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "put": {
        "summary": "Update patient details",
        "description": "Updates an existing patient's details.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Patient ID (MongoDB ObjectId)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "uhid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "age": {
            "type": "integer"
          },
          "address": {
            "type": "string"
          },
          "bloodGroup": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "emergencyContactNumber": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatientInput": {
        "type": "object",
        "required": [
          "name",
          "mobile",
          "gender",
          "address"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "mobile": {
            "type": "string",
            "minLength": 10,
            "pattern": "^\\d+$"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "dateOfBirth": {
            "type": "string"
          },
          "age": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "minLength": 3
          },
          "bloodGroup": {
            "type": "string",
            "enum": [
              "A_positive",
              "A_negative",
              "B_positive",
              "B_negative",
              "AB_positive",
              "AB_negative",
              "O_positive",
              "O_negative",
              ""
            ]
          },
          "emergencyContact": {
            "type": "string"
          },
          "emergencyContactNumber": {
            "type": "string"
          }
        }
      },
      "PatientDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Patient"
          },
          {
            "type": "object",
            "properties": {
              "appointments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Appointment"
                }
              },
              "consultations": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Consultation"
                }
              },
              "admissions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Admission"
                }
              },
              "labOrders": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LabOrder"
                }
              },
              "invoices": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "time": {
            "type": "string"
          },
          "consultationType": {
            "type": "string",
            "enum": [
              "new",
              "follow_up"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "waiting",
              "in_progress",
              "completed",
              "cancelled"
            ]
          },
          "appointmentNumber": {
            "type": "string"
          },
          "tokenNumber": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "doctor": {
            "$ref": "#/components/schemas/Doctor"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          }
        }
      },
...
      "Consultation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "chiefComplaint": {
            "type": "string"
          },
          "symptoms": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "bloodPressure": {
            "type": "string"
          },
          "pulse": {
            "type": "integer"
          },
          "respiratoryRate": {
            "type": "integer"
          },
          "oxygenSaturation": {
            "type": "integer"
          },
          "weight": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "clinicalNotes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Admission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "wardId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "bedId": {
            "type": "string"
          },
          "admissionNumber": {
            "type": "string"
          },
          "admissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "dischargeDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "admitted",
              "discharged"
            ]
          },
          "dischargeSummary": {
            "type": "string"
          },
          "finalDiagnosis": {
            "type": "string"
          },
          "followUpDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "OPD",
              "IPD"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid",
              "cancelled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number"
                },
                "total": {
                  "type": "number"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "method": {
                  "type": "string"
                },
                "transactionId": {
                  "type": "string"
                },
                "paidAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "receivedBy": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "LabOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "consultationId": {
            "type": "string"
          },
          "doctorId": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed"
            ]
          },
          "orderedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "hospital_admin",
              "receptionist",
              "doctor",
              "nurse",
              "lab_technician",
              "pharmacist",
              "billing_staff"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "Doctor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "qualification": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          }
        }
      },
...
    },
    "responses": {
      "400ValidationError": {
        "description": "Validation failed / Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "500InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}